% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RsquareMI.R
\name{RsquareSP}
\alias{RsquareSP}
\title{Calculate R-squared with Standardized Predictors}
\usage{
RsquareSP(dataset, model, beta = FALSE, cor = FALSE)
}
\arguments{
\item{dataset}{A multiply imputed dataset of class `mids` from the `mice` package.}

\item{model}{A linear model formula specifying the model to be fitted.}

\item{beta}{Logical. If `TRUE`, the function returns the standardized regression coefficients. Default is `FALSE`.}

\item{cor}{Logical. If `TRUE`, the function returns the zero-order correlations between the outcome and each predictor. Default is `FALSE`.}
}
\value{
A list of class `RsquaredMI` containing the following elements:
\item{R_squared}{The R-squared value calculated using standardized predictors.}
\item{beta}{The standardized regression coefficients (if `beta = TRUE`).}
\item{zero}{The zero-order correlations between the outcome and each predictor (if `cor = TRUE`).}
}
\description{
This function calculates the R-squared value for a linear model applied to a multiply imputed dataset.
Optionally, it can also return the standardized regression coefficients and zero-order correlations.
}
\details{
The function first completes the imputed datasets using `mice::complete`.
It then calculates the linear model on each imputed dataset and averages the standardized coefficients and correlations across imputations.
The final R-squared value is computed as the sum of the products of the averaged standardized coefficients and averaged correlations.
}
\examples{
\dontrun{
# Assuming `imp` is a multiply imputed dataset from `mice`
model <- as.formula(y ~ x1 + x2)
result <- RsquareSP(dataset = imp, model = model, beta = TRUE, cor = TRUE)
print(result$R_squared)
print(result$beta)
print(result$zero)
}

}
